refactor(bspline): update internal library imports for subdivided modules
